fix: circular dependency
Problem:

The getStartedCommand is now used in multiple
places outside of the explorer. This was causing
a circular dependency.

Solution:

Move the getStartedCommand to the utils module
