+++ src/auth/ui/vue/show.ts
@@ -23,7 +23,7 @@ import { getLogger } from '../../../shared/logger'
 import { AuthUtil as CodeWhispererAuth } from '../../../codewhisperer/util/authUtil'
 import { awsIdSignIn } from '../../../codewhisperer/util/showSsoPrompt'
 import { CodeCatalystAuthenticationProvider } from '../../../codecatalyst/auth'
-import { getStartedCommand } from '../../../codecatalyst/explorer'
+import { getStartedCommand } from '../../../codecatalyst/utils'
 import { ToolkitError } from '../../../shared/errors'
 import { isBuilderIdConnection } from '../../connection'
 import { tryAddCredentials, signout, showRegionPrompter, addConnection } from '../../utils'
